Added FastChessService, benchmarks and changed result type to long

+++ katas/2022-04-27_Strange_Chessboard/solutions/domenic_urban/java-springboot/src/main/java/com/github/irgend42/weirdchess/controller/WeirdChessController.java
@@ -17,7 +17,7 @@ public class WeirdChessController {
     private final WeirdChessService chessService;
 
     @GetMapping("calculateFieldSizes")
-    ResponseEntity<Integer[]> calculateFieldSizes(@RequestParam List<Integer> cs, @RequestParam List<Integer> rs){
+    ResponseEntity<Long[]> calculateFieldSizes(@RequestParam List<Integer> cs, @RequestParam List<Integer> rs){
         return ResponseEntity.ok(chessService.calculatePane(cs, rs));
     }
 }
