+++ katas/2022-04-27_Strange_Chessboard/solutions/domenic_urban/java-springboot/src/main/java/com/github/irgend42/weirdchess/service/WeirdChessService.java
@@ -3,5 +3,5 @@ package com.github.irgend42.weirdchess.service;
 import java.util.List;
 
 public interface WeirdChessService {
-    Integer[] calculatePane(List<Integer> cs, List<Integer> ws);
+    Long[] calculatePane(List<Integer> cs, List<Integer> ws);
 }
